% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{add_match}
\alias{add_match}
\title{Add match}
\usage{
add_match(model, match)
}
\arguments{
\item{model}{afelo_model object}

\item{match}{aflelo_match object}
}
\value{
aflelo_model with ratings updated based on the match
}
\description{
Add the results of a match to an AFLELO Model
}
\examples{
model <- aflelo_model()
match <- aflelo_match(2000, "R1", "Melbourne", "Adelaide", "M.C.G.",
                      10, 110, 102)
aflelo:::add_match(model, match)
}
