% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_team}
\alias{add_team}
\title{Add team}
\usage{
add_team(model, team)
}
\arguments{
\item{model}{aflelo_model to add team to}

\item{team}{name of the new team to add}
}
\value{
aflelo_model with new team
}
\description{
Add a new team to an AFLELO Model
}
\examples{
model <- aflelo_model()
aflelo:::add_team(model, "Gold Coast")
}
