% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{aflelo_match}
\alias{aflelo_match}
\title{AFLELO Match}
\usage{
aflelo_match(season, round, home, away, ground, pred_margin, home_total,
  away_total)
}
\arguments{
\item{season}{Integer giving season of the match}

\item{round}{String giving the round of the match}

\item{home}{String giving the name of the home team}

\item{away}{String giving the name of the away team}

\item{ground}{String giving the name of the ground}

\item{pred_margin}{Number giving the predicted margin}

\item{home_total}{Number giving the total score for the home team}

\item{away_total}{Number giving the total score for the away team}
}
\value{
A valid aflelo_match object
}
\description{
Object for storing information about AFL matches
}
\examples{
aflelo_match(2000, "R1", "Melbourne", "Adelaide", "M.C.G.", 10, 110, 102)

}
