% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{aflelo_model}
\alias{aflelo_model}
\title{AFLELO Model}
\usage{
aflelo_model(params = new_aflelo_params(), ...)
}
\arguments{
\item{params}{aflelo_params object containing parameters for the model}

\item{...}{Any other parts of the model to set}
}
\value{
A valid aflelo_model object
}
\description{
Object for storing information about the AFLELO Model
}
\examples{
model <- aflelo_model()

}
