% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{aflelo_params}
\alias{aflelo_params}
\title{AFLELO Params}
\usage{
aflelo_params(new_team_rating = 1090, new_season_adjustment = 0.1,
  hga_alpha = 6, hga_beta = 15, pred_m = 400, pred_p = 0.0464,
  adjust_k_early = 82, adjust_k_normal = 62, adjust_k_finals = 72,
  sim_sigma = 37.1)
}
\arguments{
\item{new_team_rating}{start rating for new teams}

\item{new_season_adjustment}{amount to regress at beginning of new season,
in range 0-1}

\item{hga_alpha}{alpha parameter for calculating home ground advantage,
controls the weighting given to travel}

\item{hga_beta}{beta parameter for calculating home ground advantage,
controls the weighting given to ground experience}

\item{pred_m}{m parameter for predicting results, controls how difference in
rating effects predicted result}

\item{pred_p}{p parameter for predicting margins, controls how predicted
results are converted to margins}

\item{adjust_k_early}{k value for early rounds, bigger values cause bigger
changes in ratings}

\item{adjust_k_normal}{k value for normal rounds}

\item{adjust_k_finals}{k value for finals}

\item{sim_sigma}{standard deviation used when simulating margins}
}
\value{
A valid aflelo_params object
}
\description{
Object for storing parameters for an AFLELO Model
}
\examples{
aflelo_params()

}
