% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{breed_params}
\alias{breed_params}
\title{Breed AFLELO Params}
\usage{
breed_params(population, fitness, n)
}
\arguments{
\item{population}{list of aflelo_params}

\item{fitness}{vector giving the fitness of each aflelo_params}

\item{n}{size of new population}
}
\value{
list of alfelo_params
}
\description{
Create a new population of AFLELO Params through crossover of an existing
population
}
\examples{
population <- aflelo:::new_population(5)
aflelo:::breed_params(population, rep(0.5, 5), 5)
}
