% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{calc_hga}
\alias{calc_hga}
\title{Calculate Home Ground Advantage}
\usage{
calc_hga(model, home, away, ground)
}
\arguments{
\item{model}{aflelo_model object}

\item{home}{name of the home team}

\item{away}{name of the away team}

\item{ground}{name of the ground}
}
\value{
Value giving the advantage of the home team in ratings points
}
\description{
Calculate the Home Ground Advantage between two teams playing at a particular
ground
}
\examples{
model <- aflelo_model()
aflelo:::calc_hga(model, "West Coast", "Collingwood", "Subiaco")
}
