% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{calc_new_rating}
\alias{calc_new_rating}
\title{Calculate new rating}
\usage{
calc_new_rating(model, team, real_result, pred_result)
}
\arguments{
\item{model}{aflelo_model object}

\item{team}{name of team to calculate new rating for}

\item{real_result}{real result probability}

\item{pred_result}{predicted result probability}
}
\value{
New rating value for the team
}
\description{
Calculate new rating for a team based on predicted and real results
}
\examples{
model <- aflelo_model()
match <- aflelo_match(2000, "R1", "Melbourne", "Adelaide", "M.C.G.",
                      10, 110, 102)
real_result <- aflelo:::convert_margin(model, match$real_margin)
pred_result <- aflelo:::convert_margin(model, match$pred_margin)
aflelo:::calc_new_rating(model, "Melbourne", real_result, pred_result)
}
