% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{convert_margin}
\alias{convert_margin}
\title{Convert margin}
\usage{
convert_margin(model, margin)
}
\arguments{
\item{model}{aflelo_model object}

\item{margin}{real margin for the home team}
}
\value{
Probability of the home team winning
}
\description{
Convert a real margin to a result probabilty
}
\examples{
model <- aflelo_model()
aflelo:::convert_margin(model, 30)
}
