% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{distances}
\alias{distances}
\title{Distances}
\format{\code{distances} is a matrix containing the great circle distance in
kilometres between the home city of AFL teams and the city of AFL grounds}
\usage{
distances
}
\description{
Distances between teams and grounds
}
\keyword{datasets}
