% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{evaluate_population}
\alias{evaluate_population}
\title{Evaluate population}
\usage{
evaluate_population(population, matches, start_eval = 2000, end_eval = 2016,
  pred_weight = 0.8, n_cores = 1)
}
\arguments{
\item{population}{list of aflelo_params to evaluate}

\item{matches}{data.frame of matches for training and evaluation}

\item{start_eval}{start season for calculating fitness}

\item{end_eval}{end season for calculating fitness}

\item{pred_weight}{weight given to prediction accuracy when calculating
fitness}

\item{n_cores}{number of cores to use}
}
\value{
data.frame giving the fitness of each aflelo_params
}
\description{
Evaluate a population of AFLELO Params object
}
\examples{
date("matches")
population <- aflelo::new_population(5)
evaluate_population(population, matches, start_eval = 1997, end_eval = 1998)
}
