% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{generate_params}
\alias{generate_params}
\title{Generate parameters}
\usage{
generate_params()
}
\value{
afelo_params object
}
\description{
Randomly generate a single AFLELO Params
}
\examples{
aflelo::generate_params()
}
