% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{mutate_params}
\alias{mutate_params}
\title{Muatate AFLELO Params}
\usage{
mutate_params(params, prob = 0.2)
}
\arguments{
\item{params}{aflelo_params object to mutate}

\item{prob}{probability that an individual parameter is mutated}
}
\value{
mutated aflelo_params
}
\description{
Mutate AFLELO Params
}
\examples{
params <- aflelo_params()
aflelo:::mutate_params(params)
}
