% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{new_aflelo_match}
\alias{new_aflelo_match}
\title{New AFLELO Match}
\usage{
new_aflelo_match(season, round, home, away, ground, pred_margin, real_margin,
  home_total, away_total)
}
\arguments{
\item{season}{Integer giving season of the match}

\item{round}{String giving the round of the match}

\item{home}{String giving the name of the home team}

\item{away}{String giving the name of the away team}

\item{ground}{String giving the name of the ground}

\item{pred_margin}{Number giving the predicted margin}

\item{real_margin}{Number giving the real margin}

\item{home_total}{Number giving the total score for the home team}

\item{away_total}{Number giving the total score for the away team}
}
\value{
An aflelo_match object
}
\description{
Create a new AFLELO Match
}
