% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{new_aflelo_model}
\alias{new_aflelo_model}
\title{New AFLELO Model}
\usage{
new_aflelo_model(params = new_aflelo_params())
}
\arguments{
\item{params}{aflelo_params object containing parameters for the model}
}
\value{
An aflelo_model object
}
\description{
Create a new AFLELO Model
}
