% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{new_population}
\alias{new_population}
\title{New population}
\usage{
new_population(n)
}
\arguments{
\item{n}{size of the population}
}
\value{
list of aflelo_params objects
}
\description{
Randomly generate a population of AFLELO Params
}
\examples{
population <- aflelo:::new_population(5)
}
