% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_margin}
\alias{predict_margin}
\title{Predict margin}
\usage{
predict_margin(model, pred_result)
}
\arguments{
\item{model}{aflelo_model object}

\item{pred_result}{probablilty of home win}
}
\value{
Margin for the home team
}
\description{
Convert a predicted result to a predicted margin
}
\examples{
model <- aflelo_model()
pred_result <- aflelo:::predict_result(model, "Brisbane", "Western Bulldogs",
                        "Gabba")
aflelo:::predict_margin(model, pred_result)
}
