% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_match}
\alias{predict_match}
\title{Predict match}
\usage{
predict_match(model, home, away, ground)
}
\arguments{
\item{model}{afelo_model object}

\item{home}{name of the home team}

\item{away}{name of the away team}

\item{ground}{name of the ground}
}
\value{
Probability of victory and margin for the home team
}
\description{
Predict the result and margin of a match between two teams at a particular
ground
}
\examples{
model <- aflelo_model()
predict_match(model, "Brisbane", "Western Bulldogs", "Gabba")

}
