% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_result}
\alias{predict_result}
\title{Predict result}
\usage{
predict_result(model, home, away, ground)
}
\arguments{
\item{model}{afelo_model object}

\item{home}{name of the home team}

\item{away}{name of the away team}

\item{ground}{name of the ground}
}
\value{
Probability of victory for the home team
}
\description{
Predict the result of a match between two teams at a particular ground
}
\examples{
model <- aflelo_model()
aflelo:::predict_result(model, "Brisbane", "Western Bulldogs", "Gabba")
}
