% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{print.aflelo_match}
\alias{print.aflelo_match}
\title{Print AFLELO Match}
\usage{
\method{print}{aflelo_match}(x, ...)
}
\arguments{
\item{x}{aflelo_match object to print}

\item{...}{additional arguments, not used}
}
\value{
Prints object
}
\description{
Print an AFLELO Match
}
