% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.aflelo_model}
\alias{print.aflelo_model}
\title{Print AFLELO Model}
\usage{
\method{print}{aflelo_model}(x, ...)
}
\arguments{
\item{x}{aflelo_model object to print}

\item{...}{additional arguments, not used}
}
\value{
Prints object
}
\description{
Print an AFLELO Model
}
