% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{print.aflelo_params}
\alias{print.aflelo_params}
\title{Print AFLELO Params}
\usage{
\method{print}{aflelo_params}(x, compact = FALSE, ...)
}
\arguments{
\item{x}{aflelo_params to print}

\item{compact}{whether to print in compact format}

\item{...}{additional arguments, not used}
}
\value{
Prints object
}
\description{
Print an AFLELO Params
}
