% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_matches}
\alias{simulate_matches}
\title{Simulate matches}
\usage{
simulate_matches(model, matches, n = 10000, n_cores = 1, seed = 1)
}
\arguments{
\item{model}{trained aflelo_model object}

\item{matches}{data.frame containg matches to simulate, see \link{matches} for an
example}

\item{n}{number of times to run the simulation}

\item{n_cores}{number of cores to use}

\item{seed}{random seed set before beginning simulations}
}
\value{
Ratings for teams at the end of each simulation
}
\description{
Simulate a series of AFL matches
}
\examples{
data("matches")
matches1997 <- matches[matches$Season == 1997, ]
model <- aflelo_model()
simulate_matches(model, matches1997, n = 1)

}
