% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{squeeze_value}
\alias{squeeze_value}
\title{Squeeze value}
\usage{
squeeze_value(value, lower, upper)
}
\arguments{
\item{value}{value to squeeze}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
squeezed value
}
\description{
Force a value to be within a range
}
\examples{
aflelo::squeeze_value(10, 0, 10)
aflelo::squeeze_value(10, 0, 5)
aflelo::squeeze_value(10, 15, 20)
}
