% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_model}
\alias{train_model}
\title{Traing AFLELO Model}
\usage{
train_model(model, matches)
}
\arguments{
\item{model}{aflelo_model object}

\item{matches}{data.frame containg matche results to train using, see
\link{matches} for an example}
}
\value{
aflelo_model with ratings history based on the provided matches
}
\description{
Train an AFLELO Model
}
\examples{
data("matches")
matches1997 <- matches[matches$Season == 1997, ]
model <- aflelo_model()
train_model(model, matches1997)

}
