% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update_ladder}
\alias{update_ladder}
\title{Update ladder}
\usage{
update_ladder(model, team, points, pts_for, pts_against)
}
\arguments{
\item{model}{aflelo_model to update}

\item{team}{name of team to update}

\item{points}{number of premiership points to add}

\item{pts_for}{number of points scored by the team}

\item{pts_against}{number of points scored against the team}
}
\value{
afelo_object with updated ladder
}
\description{
Update ladder information for a single team in an AFLELO Model
}
\examples{
model <- aflelo_model()
aflelo:::update_ladder(model, "Richmond", 4, 110, 100)
}
