% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update_rating_history}
\alias{update_rating_history}
\title{Update rating history}
\usage{
update_rating_history(model)
}
\arguments{
\item{model}{aflelo_model to update}
}
\value{
aflelo_model with updated rating history
}
\description{
Add the current ratings to the rating history
}
\examples{
model <- aflelo_model()
aflelo:::update_rating_history(model)
}
