% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update_ratings}
\alias{update_ratings}
\title{Update ratings}
\usage{
update_ratings(model, new_ratings)
}
\arguments{
\item{model}{aflelo_model to update}

\item{new_ratings}{vector of new ratings}
}
\value{
aflelo_model with updated ratings
}
\description{
Update the ratings for teams in an AFLELO Model
}
\examples{
model <- aflelo_model()
aflelo:::update_ratings(model, rep(1600, 16))
}
