% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{validate_aflelo_match}
\alias{validate_aflelo_match}
\title{Validate AFLELO Match}
\usage{
validate_aflelo_match(match)
}
\arguments{
\item{match}{The aflelo_match to validate}
}
\value{
Raises an error if invalid, otherwise returns the aflelo_match
}
\description{
Validate an AFLELO Match object
}
