% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{validate_aflelo_model}
\alias{validate_aflelo_model}
\title{Validate AFLELO Model}
\usage{
validate_aflelo_model(model)
}
\arguments{
\item{model}{The aflelo_model to validate}
}
\value{
Raises an error if invalid, otherwise returns the aflelo_model
}
\description{
Validate an AFLELO Model object
}
