% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{validate_aflelo_params}
\alias{validate_aflelo_params}
\title{Validate AFLELO Params}
\usage{
validate_aflelo_params(params)
}
\arguments{
\item{params}{aflelo_params to validate}
}
\value{
Raises an error if invalid, otherwise returns the aflelo_params
}
\description{
Validate and AFLELO Params object
}
