% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{matches}
\alias{matches}
\title{Matches}
\format{\code{matches} is a data.frame containing the details of AFL matches from
the 1997 season to the end of the 2017 seaons, obtained from afltables.com}
\usage{
matches
}
\description{
Details of AFL matches
}
\details{
Some changes have been made to what is available on afltables.com:
\itemize{
\item A season column has been added
\item Scores have been split into goals, behinds and totals columns
\item "Brisbane Lions" have been renamed to "Brisbane"
\item "Kangaroos" have been renamed to "North Melbourne"
\item "Western Bulldog" have been renamed to "Western Bulldogs"
}
}
\keyword{datasets}
