% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update_rating}
\alias{update_rating}
\title{Update rating}
\usage{
update_rating(model, team, new_rating, points, pts_for, pts_against)
}
\arguments{
\item{model}{aflelo_model to update}

\item{team}{name of team to update}

\item{new_rating}{new rating value}
}
\value{
afelo_object with updated rating
}
\description{
Update rating for a single team in an AFLELO Model
}
\examples{
model <- aflelo_model()
aflelo:::update_rating(model, "Richmond", 1600)
}
